# [4.x] Improve comb search driver performance