@JohnathonKoster 2023-12-01
Doing some quick tests, this has also improved things dramatically on a site I've been working on with a large amount of index entries